chore(cli): Start command cleanup
